# Add a smaller default config option for decode