@drisspg 2024-08-16
@joydddd @Chillee  The failure looks real, not sure why halving the block size hurt the numerics so much but probs something is broken